[+] Add of the hiden API key and Kernel install
Now when you run the programme it will check if you have all the necessary kernel to run the process. If not it will install them for you.
You API key is now secret and hiden if you want to see it just click on the buton next to your key.